Allow `make tests` from inside of the src/ directory
For now, the Blacksmith package is a source submodule.  It's not the
greatest, but it works for fast iteration of the dev process, esp. given
that we need a BOSH environment to adequately test this thing.

Now, from inside of that submodule, if you set your GOPATH to
`$reporoot/src/blacksmith:$GOPATH`, you can run tests.  Yay!